Move fecth related functions to fechingData.js
